+++ r-kvstore-20150101/alibabacloud_r_kvstore20150101/models/_describe_accounts_request.py
@@ -11,20 +11,29 @@ class DescribeAccountsRequest(DaraModel):
         instance_id: str = None,
         owner_account: str = None,
         owner_id: int = None,
+        page_number: int = None,
+        page_size: int = None,
         resource_owner_account: str = None,
         resource_owner_id: int = None,
+        search_account_name: str = None,
         security_token: str = None,
     ):
-        # The name of the account that you want to query.
+        # The name of the account to query.
         self.account_name = account_name
-        # The ID of the instance.
+        # The instance ID.
         # 
         # This parameter is required.
         self.instance_id = instance_id
         self.owner_account = owner_account
         self.owner_id = owner_id
+        # The page number.
+        self.page_number = page_number
+        # The number of entries per page.
+        self.page_size = page_size
         self.resource_owner_account = resource_owner_account
         self.resource_owner_id = resource_owner_id
+        # The fuzzy match string for the account name.
+        self.search_account_name = search_account_name
         self.security_token = security_token
 
     def validate(self):
@@ -47,11 +56,20 @@ class DescribeAccountsRequest(DaraModel):
         if self.owner_id is not None:
             result['OwnerId'] = self.owner_id
 
+        if self.page_number is not None:
+            result['PageNumber'] = self.page_number
+
+        if self.page_size is not None:
+            result['PageSize'] = self.page_size
+
         if self.resource_owner_account is not None:
             result['ResourceOwnerAccount'] = self.resource_owner_account
 
         if self.resource_owner_id is not None:
             result['ResourceOwnerId'] = self.resource_owner_id
+
+        if self.search_account_name is not None:
+            result['SearchAccountName'] = self.search_account_name
 
         if self.security_token is not None:
             result['SecurityToken'] = self.security_token
@@ -72,11 +90,20 @@ class DescribeAccountsRequest(DaraModel):
         if m.get('OwnerId') is not None:
             self.owner_id = m.get('OwnerId')
 
+        if m.get('PageNumber') is not None:
+            self.page_number = m.get('PageNumber')
+
+        if m.get('PageSize') is not None:
+            self.page_size = m.get('PageSize')
+
         if m.get('ResourceOwnerAccount') is not None:
             self.resource_owner_account = m.get('ResourceOwnerAccount')
 
         if m.get('ResourceOwnerId') is not None:
             self.resource_owner_id = m.get('ResourceOwnerId')
+
+        if m.get('SearchAccountName') is not None:
+            self.search_account_name = m.get('SearchAccountName')
 
         if m.get('SecurityToken') is not None:
             self.security_token = m.get('SecurityToken')