+++ r-kvstore-20150101/alibabacloud_r_kvstore20150101/models/_describe_parameter_templates_response_body.py
@@ -16,14 +16,14 @@ class DescribeParameterTemplatesResponseBody(DaraModel):
         parameters: main_models.DescribeParameterTemplatesResponseBodyParameters = None,
         request_id: str = None,
     ):
-        # The database engine that is run on the instance. The value **Redis** is returned for this parameter.
+        # The database type. The return value is fixed as **Redis**.
         self.engine = engine
         # The major version of the instance.
         self.engine_version = engine_version
-        # The number of parameters that are supported by the instance.
+        # The number of parameters.
         self.parameter_count = parameter_count
         self.parameters = parameters
-        # The ID of the request.
+        # The request ID.
         self.request_id = request_id
 
     def validate(self):